Move funcoes executadas no inicio do carregamento do plugin
